cartelet v0.1.0 Cartelet.Math.Derivative

Provides functions for evaluating derivatives of univariate functions.

Functions

backward_difference(f, x, h)

backward_difference((float() -> float()), float(), float()) :: float()

Evaluates the first derivative of a univariate function f at the value x, with step size h, using a backward difference scheme.

backward_difference_second(f, x, h)

backward_difference_second((float() -> float()), float(), float()) :: float()

Evaluates the second derivative of a univariate function f at the value x, with step size h, using a backward difference scheme.

central_difference(f, x, h)

central_difference((float() -> float()), float(), float()) :: float()

Evaluates the first derivative of a univariate function f at the value x, with step size h, using a central difference scheme.

central_difference_second(f, x, h)

central_difference_second((float() -> float()), float(), float()) :: float()

Evaluates the second derivative of a univariate function f at the value x, with step size h, using a central difference scheme.

forward_difference(f, x, h)

forward_difference((float() -> float()), float(), float()) :: float()

Evaluates the first derivative of a univariate function f at the value x, with step size h, using a forward difference scheme.

forward_difference_second(f, x, h)

forward_difference_second((float() -> float()), float(), float()) :: float()

Evaluates the second derivative of a univariate function f at the value x, with step size h, using a forward difference scheme.