cartelet v0.1.0 Cartelet.Math.Integral

Provides functions for evaluating integrals of univariate functions.

Functions

simpson_13(f, a, b)

simpson_13((float() -> float()), float(), float()) :: float()

Integrates a univariate f between a and b using Simpson’s 1/3 rule.

simpson_13_composite(f, a, b, n)

simpson_13_composite((float() -> float()), float(), float(), integer()) ::
  float()

Integrates a univariate f between a and b using Simpson’s composite 1/3 rule with n segments.

simpson_38(f, a, b)

simpson_38((float() -> float()), float(), float()) :: float()

Integrates a univariate f between a and b using Simpson’s 3/8 rule.

trapezoid(f, a, b)

trapezoid((float() -> float()), float(), float()) :: float()

Integrates a univariate function f between a and b using the trapezoidal rule.

trapezoid_composite(f, a, b, n)

trapezoid_composite((float() -> float()), float(), float(), integer()) ::
  float()

Integrates a univariate function f between a and b using the composite trapezoidal rule with n segments.