cartelet v0.1.0 Cartelet.Math.OdeInt

Provides functions for numerical integration of systems of ordinary differential equations.

(N.B.: “System” in this context can refer to either systems of ODEs or the system itself that we are integrating. We’ll stick with this slightly confusing terminology until I come up with something better…)

Types

ode_func()

ode_func() :: (float(), state(), keyword() -> state())

Evaluates a system of ODEs on a state at a given point in time.

state()

state() :: %Numerix.Tensor{dims: term(), items: term(), shape: term()}

Represents the state of a system of any dimension.

Functions

integrate(ode, state_init, dt, steps, method \\ :rk4, params \\ [])

integrate(ode_func(), state(), float(), integer(), atom(), keyword()) ::
  {:ok, float(), state()}

Integrates the given ODE system with the given initial conditions and step size, over a number of steps.

If the ODE parameters and integration method are not specified, defaults to pre-defined parameters and 4th-order Runge-Kutta.

Returns the new state.

lorenz(t, state, params)

lorenz(float(), state(), keyword()) :: state()

Evaluates the Lorenz system (a.k.a. Lorenz attractor).

Returns the new state.

Reference: Wikipedia

rk4(ode, t, state, dt, params)

rk4(ode_func(), float(), state(), float(), keyword()) :: {float(), state()}

Integrates an ODE system over a single step, at a given point in time, via the 4th-order Runge-Kutta method (RK4).

Reference: Wikipedia